+++ pytriton/proxy/communication.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2022-2023, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@ import dataclasses
 import json
 import struct
 from multiprocessing.shared_memory import SharedMemory
-from typing import Callable, Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Literal, Optional, Tuple, Union
 
 import numpy as np
 
@@ -28,7 +28,7 @@ import numpy as np
 # https://github.com/triton-inference-server/python_backend/blob/main/src/resources/triton_python_backend_utils.py
 
 
-def _serialize_byte_tensor(input_tensor):
+def _serialize_byte_tensor(tensor) -> bytes:
     """Serializes a bytes tensor into a flat numpy array of length prepended bytes.
 
     The numpy array should use dtype of np.object_. For np.bytes_,
@@ -44,44 +44,42 @@ def _serialize_byte_tensor(input_tensor):
     Raises:
         UnicodeEncodeErrors: raised when try to cast to string of non-bytes items fails
     """
-    if input_tensor.size == 0:
-        return ()
+    if tensor.size == 0:
+        return b""
 
     # If the input is a tensor of string/bytes objects, then must flatten those
     # into a 1-dimensional array containing the 4-byte byte size followed by the
     # actual element bytes. All elements are concatenated together in "C" order.
-    if (input_tensor.dtype == np.object_) or (input_tensor.dtype.type == np.bytes_):
-        flattened_ls = []
-        total_len = 0
-        for obj in np.nditer(input_tensor, flags=["refs_ok"], order="C"):
-            # If directly passing bytes to BYTES type,
-            # don't convert it to str as Python will encode the
-            # bytes which may distort the meaning
-            if input_tensor.dtype == np.object_:
-                if type(obj.item()) == bytes:
-                    s = obj.item()
-                else:
-                    s = str(obj.item()).encode("utf-8")
-            else:
-                s = obj.item()
-            item_len = len(s)
-            flattened_ls.append(struct.pack("<I", item_len))
-            flattened_ls.append(s)
-            total_len += 4 + item_len
-        flattened_ls.insert(0, struct.pack("<I", total_len))
-        flattened = b"".join(flattened_ls)
-        return flattened
-    return None
+    assert (tensor.dtype == np.object_) or (tensor.dtype.type == np.bytes_)
+    flattened_ls = []
+    total_len = 0
+    for obj in np.nditer(tensor, flags=["refs_ok"], order="C"):
+        # If directly passing bytes to BYTES type,
+        # don't convert it to str as Python will encode the
+        # bytes which may distort the meaning
+        if tensor.dtype == np.object_ and type(obj.item()) != bytes:
+            s = str(obj.item()).encode("utf-8")
+        else:
+            s = obj.item()
+        item_len = len(s)
+        flattened_ls.append(struct.pack("<I", item_len))
+        flattened_ls.append(s)
+        total_len += struct.calcsize("<I") + item_len
+    flattened_ls.insert(0, struct.pack("<I", total_len))
+    flattened = b"".join(flattened_ls)
+    return flattened
 
 
 # copy from
 # https://github.com/triton-inference-server/python_backend/blob/main/src/resources/triton_python_backend_utils.py
-def _deserialize_bytes_tensor(encoded_tensor, dtype):
+def _deserialize_bytes_tensor(encoded_tensor, dtype, order: Literal["C", "F"] = "C") -> np.ndarray:
     """Deserializes an encoded bytes tensor into an numpy array of dtype of python objects.
 
     Args:
         encoded_tensor : The encoded bytes tensor where each element has its length in
         first 4 bytes followed by the content
+        dtype: The dtype of the numpy array to deserialize to.
+        order: The order of the numpy array to deserialize to.
 
     Returns:
     The 1-D numpy array of type object containing the deserialized bytes in 'C' order.
@@ -97,101 +95,202 @@ def _deserialize_bytes_tensor(encoded_tensor, dtype):
         item = struct.unpack_from(f"<{item_length}s", val_buf, offset)[0]
         offset += item_length
         strs.append(item)
-    return np.array(strs, dtype=dtype)
+    return np.array(strs, dtype=dtype, order=order)
 
 
-@dataclasses.dataclass
-class TensorInfo:
-    """Data class for storing numpy array schema and buffer information."""
+_MAX_DTYPE_DESCR = 8
+_PARTIAL_HEADER_FORMAT = f"<{_MAX_DTYPE_DESCR}scH"
 
-    memory_range: Tuple[int, int]
-    shape: Tuple[int, ...]
-    dtype: str
+
+def _pack_header(shape: Tuple[int, ...], dtype: np.dtype, order: Literal["C", "F"] = "C") -> bytes:
+    header_format = _PARTIAL_HEADER_FORMAT + "Q" * len(shape)
+    dtype_descr = np.lib.format.dtype_to_descr(dtype)
+    assert (
+        len(dtype_descr) <= _MAX_DTYPE_DESCR
+    ), f"dtype descr is too long; dtype_descr={dtype_descr} max={_MAX_DTYPE_DESCR}"
+    return struct.pack(header_format, dtype_descr.encode("utf-8"), order.encode("ascii"), len(shape), *shape)
+
+
+def _unpack_header(header: bytes) -> Tuple[Tuple[int, ...], np.dtype, Literal["C", "F"]]:
+    shape_offset = struct.calcsize(_PARTIAL_HEADER_FORMAT)
+    dtype_descr, order, ndim = struct.unpack_from(_PARTIAL_HEADER_FORMAT, header, offset=0)
+    shape = struct.unpack_from("Q" * ndim, header, offset=shape_offset)
+    dtype = np.lib.format.descr_to_dtype(dtype_descr.decode("utf-8").rstrip("\x00"))
+    order = order.decode("ascii")
+    return shape, dtype, order
+
+
+def serialize_numpy_with_struct_header(tensor: np.ndarray) -> List[Union[bytes, memoryview]]:
+    """Serialize numpy array to list of bytes and memoryviews.
+
+    Args:
+        tensor: numpy array to serialize
+
+    Returns:
+        List of data frames in form of bytes and memoryviews
+    """
+    if tensor.dtype.hasobject:
+        data = _serialize_byte_tensor(tensor.ravel())
+        order = "C"  # as _serialize_byte_tensor returns C-ordered array
+    else:
+        if not tensor.data.contiguous:
+            tensor = np.ascontiguousarray(tensor)
+        data = tensor.data
+        order = "C" if tensor.flags.c_contiguous else "F"
+
+    header = _pack_header(tensor.shape, tensor.dtype, order)
+    frames = [header, data]
+    return frames
+
+
+def deserialize_numpy_with_struct_header(frames: List[Union[bytes, memoryview]]) -> np.ndarray:
+    """Deserialize numpy array from list of bytes and memoryviews.
+
+    Args:
+        frames: List of data frames in form of bytes and memoryviews
+
+    Returns:
+        numpy array
+    """
+    header, data = frames
+    shape, dtype, order = _unpack_header(header)
+    if dtype.hasobject:
+        tensor = _deserialize_bytes_tensor(data, dtype).reshape(shape)
+    else:
+        tensor = np.ndarray(shape, dtype=dtype, buffer=data, order=order)
+    return tensor
+
+
+def calc_serialized_size_of_numpy_with_struct_header(tensor: np.ndarray) -> List[int]:
+    """Calculate size of serialized numpy array.
+
+    Args:
+        tensor: numpy array to serialize
+
+    Returns:
+        List of sizes of data frames
+    """
+    header_size = struct.calcsize(_PARTIAL_HEADER_FORMAT) + struct.calcsize("Q") * len(tensor.shape)
+    if tensor.dtype.hasobject:
+        items_sizes = []
+        order = "C" if tensor.flags.c_contiguous else "F"
+        for obj in np.nditer(tensor, flags=["refs_ok"], order=order):
+            if tensor.dtype == np.object_ and type(obj.item()) != bytes:
+                s = str(obj.item()).encode("utf-8")
+            else:
+                s = obj.item()
+            items_sizes.append(len(s))
+
+        # total_size + for size of each item + each item
+        data_size = struct.calcsize("<I") + struct.calcsize("<I") * len(items_sizes) + sum(items_sizes)
+    else:
+        data_size = tensor.nbytes
+
+    return [header_size, data_size]
+
+
+@dataclasses.dataclass(frozen=True)
+class TensorId:
+    """Data class for storing id of tensor in Tensor Store."""
+
+    memory_name: str
+    memory_offset: int
+
+    @classmethod
+    def from_str(cls, tensor_id: str):
+        """Create TensorId object from string."""
+        memory_name, memory_offset = tensor_id.split(":", maxsplit=1)
+        return cls(memory_name, int(memory_offset))
+
+    def __str__(self):
+        """Convert TensorId object to string."""
+        return f"{self.memory_name}:{self.memory_offset}"
 
 
 @dataclasses.dataclass
 class MetaRequestResponse:
     """Data class for storing input/output data and parameters."""
 
-    data: Dict[str, TensorInfo]
+    data: Dict[str, TensorId]
     parameters: Optional[Dict[str, Union[str, int, bool]]] = None
 
 
 @dataclasses.dataclass
-class InferenceHandlerRequest:
+class InferenceHandlerRequests:
     """Object transferred from proxy backend to callback handler containing input data."""
 
     requests: List[MetaRequestResponse]
-    memory_name: str
 
     @classmethod
-    def from_bytes(cls, content: bytes) -> "InferenceHandlerRequest":
-        """Reconstruct Request object from bytes.
+    def from_bytes(cls, content: bytes) -> "InferenceHandlerRequests":
+        """Reconstruct InferenceHandlerRequests object from bytes.
 
         Args:
             content: bytes to parse
         """
-        request_dict = json.loads(content)
-        request_list = request_dict["requests"]
-
+        requests = json.loads(content)
         return cls(
             requests=[
                 MetaRequestResponse(
                     data={
-                        input_name: TensorInfo(**tensor_info) for input_name, tensor_info in req_dict["data"].items()
+                        input_name: TensorId.from_str(tensor_id)
+                        for input_name, tensor_id in request.get("data", {}).items()
                     },
-                    parameters=req_dict["parameters"],
+                    parameters=request.get("parameters"),
                 )
-                for req_dict in request_list
-            ],
-            memory_name=request_dict["memory_name"],
+                for request in requests["requests"]
+            ]
         )
 
     def as_bytes(self) -> bytes:
-        """Serializes Request object to bytes."""
-        request_dict = {
-            "requests": [dataclasses.asdict(req) for req in self.requests],
-            "memory_name": self.memory_name,
+        """Serializes InferenceHandlerRequests object to bytes."""
+        requests = {
+            "requests": [
+                {
+                    "data": {input_name: str(tensor_id) for input_name, tensor_id in request.data.items()},
+                    "parameters": request.parameters,
+                }
+                for request in self.requests
+            ]
         }
-        return json.dumps(request_dict).encode("utf-8")
+        return json.dumps(requests).encode("utf-8")
 
 
 @dataclasses.dataclass
-class InferenceHandlerResponse:
+class InferenceHandlerResponses:
     """Object transferred from callback handler containing output data."""
 
     responses: Optional[List[MetaRequestResponse]] = None
     error: Optional[str] = None
-    memory_name: Optional[str] = None
 
     @classmethod
-    def from_bytes(cls, content: bytes) -> "InferenceHandlerResponse":
-        """Reconstruct Response object from bytes.
+    def from_bytes(cls, content: bytes) -> "InferenceHandlerResponses":
+        """Reconstruct InferenceHandlerResponses object from bytes.
 
         Args:
             content: bytes to parse
         """
-        response_dict = json.loads(content)
-        responses = response_dict.get("responses", [])
-        memory_name = response_dict.get("memory_name", None)
+        responses = json.loads(content)
         return cls(
             responses=[
                 MetaRequestResponse(
-                    {output_name: TensorInfo(**tensor_info) for output_name, tensor_info in resp_dict.items()}
+                    {
+                        output_name: TensorId.from_str(tensor_id)
+                        for output_name, tensor_id in response.get("data", {}).items()
+                    }
                 )
-                for resp_dict in responses
+                for response in responses.get("responses", [])
             ],
-            error=response_dict.get("error"),
-            memory_name=memory_name,
+            error=responses.get("error"),
         )
 
     def as_bytes(self) -> bytes:
-        """Serializes Response object to bytes."""
-        result = {"error": self.error, "memory_name": self.memory_name}
+        """Serializes InferenceHandlerResponses object to bytes."""
+        result = {"error": self.error}
         if self.responses:
             result["responses"] = [
-                {output_name: dataclasses.asdict(tensor_info) for output_name, tensor_info in resp_dict.data.items()}
-                for resp_dict in self.responses
+                {"data": {output_name: str(tensor_id) for output_name, tensor_id in response.data.items()}}
+                for response in self.responses
             ]
         return json.dumps(result).encode("utf-8")
 
@@ -203,117 +302,116 @@ class ShmManager:
         """Initialize ShmManager class."""
         self._shm_buffer = None
         self._serialized_arrays = {}
-        self._memory_index = 0
+        self._offset = 0
+        self.serialize = serialize_numpy_with_struct_header
+        self.deserialize = deserialize_numpy_with_struct_header
+        self._calc_serialized_size = calc_serialized_size_of_numpy_with_struct_header
 
-    def _init_buffer(self, _required_buffer_size: int):
-        if self._shm_buffer and self._shm_buffer.size < _required_buffer_size:
+    def calc_serialized_size(self, tensor) -> int:
+        """Calculate size of serialized tensor.
+
+        Include frames storage (its sizes and total size).
+
+        Args:
+            tensor: numpy array to serialize
+
+        Returns:
+            size of serialized tensor
+        """
+        # frames sizes + total size + frames sizes
+        return sum(self._calc_serialized_size(tensor)) + struct.calcsize("<I") + 2 * struct.calcsize("<I")
+
+    def reset_buffer(self, required_buffer_size: int):
+        """Reset shared memory buffer.
+
+        Reallocate buffer if it is not big enough.
+
+        Args:
+            required_buffer_size: size of buffer to allocate
+        """
+        if self._shm_buffer and self._shm_buffer.size < required_buffer_size:
             self.dispose()
 
         if self._shm_buffer is None:
-            self._shm_buffer = SharedMemory(create=True, size=_required_buffer_size)
+            self._shm_buffer = SharedMemory(create=True, size=required_buffer_size)
 
-        self._memory_index = 0
+        self._offset = 0
 
-    def _calc_required_buffer(self, requests_or_responses: List) -> int:
-        self._serialized_arrays.clear()
-        required_buffer_size_sum = 0
-        for index, request_or_response in enumerate(requests_or_responses):
-            for input_name, np_array in request_or_response.data.items():
-                if np_array.dtype == np.object_ or np_array.dtype.type == np.bytes_:
-                    serialized_np_array = _serialize_byte_tensor(np_array)
-                    required_buffer_size = len(serialized_np_array)
-                    self._serialized_arrays[(index, input_name)] = serialized_np_array
-                else:
-                    required_buffer_size = np_array.nbytes
-                required_buffer_size_sum += required_buffer_size
-        return required_buffer_size_sum
-
-    def _get_buffer_for_write(self, size: int) -> Tuple[memoryview, Tuple[int, int]]:
-        buf_range = (self._memory_index, self._memory_index + size)
-        sub_buf = self._shm_buffer.buf[buf_range[0] : buf_range[1]]
-        self._memory_index += size
-        return sub_buf, buf_range
-
-    def _get_buffer_for_read(self, buf_range: Tuple[int, int]) -> memoryview:
-        sub_buf = self._shm_buffer.buf[buf_range[0] : buf_range[1]]
-        return sub_buf
-
-    def memory_name(self) -> Optional[str]:
-        """Returns name of shared memory buffer."""
-        return self._shm_buffer.name if self._shm_buffer else None
-
-    def to_shm(self, requests_or_responses: List, factory: Callable) -> List:
-        """Serialize list of requests or responses to shared memory.
+    def append(self, tensor: np.ndarray) -> TensorId:
+        """Append tensor to shared memory buffer.
 
         Args:
-            requests_or_responses: input list of requests or reponses for serialization
-            factory: function to create output request or response object
+            tensor: numpy array to append
 
         Returns:
-            corresponding structure where each numpy array is replaced with TensorInfo description
+            TensorId object
         """
-        self._serialized_arrays.clear()
-        required_buffer_size_sum = self._calc_required_buffer(requests_or_responses)
-        self._init_buffer(required_buffer_size_sum)
-
-        return [
-            factory(
-                {
-                    input_name: self._wrap_array(req_res_index, input_name, req_resp.data[input_name])
-                    for input_name in req_resp.data
-                },
-                req_resp,
+        if self._shm_buffer is None:
+            raise RuntimeError(
+                "Shared memory buffer is not initialized. Call reset_buffer(required_buffer_size) first."
             )
-            for req_res_index, req_resp in enumerate(requests_or_responses)
-        ]
 
-    def from_shm(
-        self, requests_or_responses: List[MetaRequestResponse], memory_name: str, output_factory: Callable
-    ) -> List:
-        """Deserialize list of requests or responses from shared memory.
+        frames = self.serialize(tensor)
+        offset = self._offset
+        data_copied_len = self._copy_frames(frames, offset)
+        self._offset += data_copied_len
+        tensor_id = TensorId(self._shm_buffer.name, offset)
+        return tensor_id
+
+    def get(self, tensor_id: Union[str, TensorId]) -> np.ndarray:
+        """Get tensor from shared memory buffer.
 
         Args:
-            requests_or_responses: input list of meta requests/reponses (with TensorInfo) for deserialization
-            memory_name: share memory buffer name
-            output_factory: type of return list elements
+            tensor_id: TensorId object or string representation of TensorId
 
         Returns:
-            list of request dicts deserialized from shared memory
+            numpy array
         """
+        if isinstance(tensor_id, str):
+            tensor_id = TensorId.from_str(tensor_id)
+        frames = self._handle_frames(tensor_id.memory_name, tensor_id.memory_offset)
+        tensor = self.deserialize(frames)
+        return tensor
+
+    def _copy_frames(self, frames: List[Union[bytes, memoryview]], offset) -> int:
+        # caller should ensure that self._shm_buffer is initialized
+        shm_buffer: SharedMemory = self._shm_buffer  # type: ignore
+
+        total_size = struct.calcsize("<I")  # start after total_size; max 4GB for all frames
+        for frame in frames:
+            if isinstance(frame, bytes):
+                frame = memoryview(frame)
+
+            assert frame.contiguous, "Only contiguous arrays are supported"
+            struct.pack_into("<I", shm_buffer.buf, offset + total_size, frame.nbytes)
+            total_size += struct.calcsize("<I")
+            shm_buffer.buf[offset + total_size : offset + total_size + frame.nbytes] = frame.cast("B")
+
+            total_size += frame.nbytes
+
+        struct.pack_into("<I", shm_buffer.buf, offset, total_size)
+        return total_size
+
+    def _handle_frames(self, memory_name, block_offset: int) -> List[memoryview]:
+
         if self._shm_buffer and self._shm_buffer.name != memory_name:
             self.dispose()
         if self._shm_buffer is None:
             self._shm_buffer = SharedMemory(memory_name, create=False)
 
-        results = [
-            output_factory({input_name: self.as_np_array(info) for input_name, info in req_resp.data.items()}, req_resp)
-            for req_resp in requests_or_responses
-        ]
-        return results
+        if self._shm_buffer.size < block_offset:
+            raise RuntimeError(f"Shared memory buffer is too small for requested offset {block_offset}")
 
-    def _wrap_array(self, req_resp_index: int, input_name: str, np_array: np.ndarray) -> TensorInfo:
-        """Copies numpy array data to shared memory.
-
-        Reuse shared memory if possible. Reallocates shared memory if needed.
-
-        Args:
-            req_resp_index: request or reponse index
-            input_name: input name
-            np_array: source numpy array
-        """
-        if np_array.dtype == np.object_ or np_array.dtype.type == np.bytes_:
-            serialized_np_array = self._serialized_arrays[
-                (req_resp_index, input_name)
-            ]  # _serialize_byte_tensor(np_array)
-            required_buffer_size = len(serialized_np_array)
-            buf, buf_range = self._get_buffer_for_write(required_buffer_size)
-            buf[:required_buffer_size] = serialized_np_array
-        else:
-            required_buffer_size = np_array.nbytes
-            buf, buf_range = self._get_buffer_for_write(required_buffer_size)
-            shm_array = np.ndarray(shape=np_array.shape, dtype=np_array.dtype, buffer=buf)
-            shm_array[:] = np_array
-        return TensorInfo(buf_range, np_array.shape, str(np_array.dtype))
+        frames = []
+        (total_size,) = struct.unpack_from("<I", self._shm_buffer.buf, block_offset)
+        offset = struct.calcsize("<I")
+        while offset < total_size:
+            (frame_size,) = struct.unpack_from("<I", self._shm_buffer.buf, block_offset + offset)
+            offset += struct.calcsize("<I")
+            frame = self._shm_buffer.buf[block_offset + offset : block_offset + offset + frame_size]
+            offset += frame_size
+            frames.append(frame)
+        return frames
 
     def dispose(self):
         """Free resources used by this wrapper."""
@@ -329,20 +427,4 @@ class ShmManager:
                 # when /dev/shm file doesn't exist - removed by other process
                 unregister(self._shm_buffer._name, "shared_memory")  # pytype: disable=attribute-error
         self._shm_buffer = None
-
-    def as_np_array(self, info: TensorInfo) -> np.ndarray:
-        """Create numpy array based on shared memory buffer.
-
-        Args:
-            info Tensor info for deserialization
-
-        Returns:
-            numpy array based on shared memory.
-        """
-        dtype = np.dtype(info.dtype)
-        buf = self._get_buffer_for_read(info.memory_range)
-
-        if dtype == np.object_ or dtype.type == np.bytes_:
-            return _deserialize_bytes_tensor(buf, dtype).reshape(info.shape)
-        else:
-            return np.ndarray(shape=info.shape, dtype=dtype, buffer=buf)
+        self._memory_name = None